voice-quiz-back/pyproject.toml
[project]
name = "voice-quiz-back"
version = "0.1.0"
description = "Voice Quiz Backend API"
requires-python = ">=3.11"
dependencies = [
    "Flask",
    "flask-cors",
    "openai",
    "numpy",
    "python-dotenv",
]

[project.scripts]
dev = "run:main"
start = "run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
]

[tool.ruff]
# Set the maximum line length to 100
line-length = 100
target-version = "py311"

# Enable auto-fixing
fix = true

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
]

[tool.ruff.lint]
# Enable recommended rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"
